fixed bugs

+++ bank-srv/src/main/java/org/statsenko/service/services/AccountService.java
@@ -1,6 +1,7 @@
 package org.statsenko.service.services;
 
 import dto.request.AccountDto;
+import dto.response.MessageResponse;
 import lombok.RequiredArgsConstructor;
 import org.mapstruct.factory.Mappers;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,24 +31,47 @@ public class AccountService {
     }
 
     @Loggable
-    public List<AccountDto> getClientAccount(int id){
-        List<AccountDto> accountDtoList = REST_MAPPER.toDtoList(accountRepository.findClientByAccount(id));
-        return accountDtoList;
+    public MessageResponse<List<AccountDto>> getClientAccount(int id){
+        MessageResponse message = new MessageResponse();
+        if (accountRepository.findClientByAccount(id).isEmpty()){
+            message.setExceptionName("Not found");
+            message.setTechnicalDescription("Client with your ID not found");
+        }
+        else {
+            List<AccountDto> accountDtoList = REST_MAPPER.toDtoList(accountRepository.findClientByAccount(id));
+            message.setResponse(accountDtoList);
+        }
+        return message;
     }
 
     @Loggable
     @Transactional(propagation = Propagation.REQUIRED)
-    public AccountDto createAccount(AccountDto accountDto){
-        Account account = REST_MAPPER.toEntity(accountDto);
-
-        accountRepository.save(account);
-
-        return accountDto;
+    public MessageResponse<AccountDto> createAccount(AccountDto accountDto){
+        MessageResponse message = new MessageResponse();
+        if (accountRepository.existsByNumberAccount(accountDto.getNumberAccount())){
+            message.setExceptionName("Similar number");
+            message.setTechnicalDescription("Bank with this number account already exist");
+        }
+        else {
+            Account account = REST_MAPPER.toEntity(accountDto);
+            accountRepository.save(account);
+            message.setResponse(accountDto);
+        }
+        return message;
     }
 
     @Loggable
     @Transactional(propagation = Propagation.REQUIRED)
-    public void deleteAccount(int id){
-        accountRepository.deleteById(id);
+    public MessageResponse deleteAccount(int id){
+        MessageResponse message = new MessageResponse();
+        if (accountRepository.findById(id).orElse(null)==null){
+            message.setExceptionName("Not found");
+            message.setTechnicalDescription("Account with your ID not found");
+        }
+        else {
+            accountRepository.deleteById(id);
+            message.setResponse("Account deleted");
+        }
+        return message;
     }
 }
