+++ bank-srv/src/main/java/org/statsenko/service/services/ProfileService.java
@@ -1,6 +1,7 @@
 package org.statsenko.service.services;
 
 import dto.request.ProfileDto;
+import dto.response.MessageResponse;
 import lombok.RequiredArgsConstructor;
 import org.mapstruct.factory.Mappers;
 import org.springframework.stereotype.Service;
@@ -23,9 +24,17 @@ public class ProfileService {
     private final ProfileRepository profileRepository;
 
     @Loggable
-    public ProfileDto getProfileById(int id){
-        ProfileDto profile = REST_MAPPER.toDto(profileRepository.getById(id));
-        return profile;
+    public MessageResponse<ProfileDto> getProfileById(int id){
+        MessageResponse message = new MessageResponse();
+        if (profileRepository.findById(id).orElse(null)==null){
+            message.setExceptionName("Not found");
+            message.setTechnicalDescription("Profile with your ID not found");
+        }
+        else {
+            ProfileDto profile = REST_MAPPER.toDto(profileRepository.getById(id));
+            message.setResponse(profile);
+        }
+        return message;
     }
 
     @Loggable
@@ -36,24 +45,49 @@ public class ProfileService {
 
     @Loggable
     @Transactional(propagation = Propagation.REQUIRED)
-    public ProfileDto createProfile(MultipartFile photo,ProfileDto profileDto) throws Exception {
-        Profile profile = REST_MAPPER.toEntity(profileDto);
-        profile.setPhoto(photo.getBytes());
-        profileRepository.save(profile);
-        return profileDto;
+    public MessageResponse<ProfileDto> createProfile(MultipartFile photo,ProfileDto profileDto) throws Exception {
+        MessageResponse message = new MessageResponse();
+        if (profileRepository.existsByLogin(profileDto.getLogin())){
+            message.setExceptionName("Similar login");
+            message.setTechnicalDescription("Bank with this login already exist");
+        }
+        else {
+            Profile profile = REST_MAPPER.toEntity(profileDto);
+            profile.setPhoto(photo.getBytes());
+            profileRepository.save(profile);
+            message.setResponse(profile);
+        }
+        return message;
     }
 
     @Loggable
     @Transactional(propagation = Propagation.REQUIRED)
-    public ProfileDto editProfile(ProfileDto profileDto, int id){
-        Profile profile = REST_MAPPER.update(profileDto, profileRepository.getById(id));
-        profileRepository.save(profile);
-        return profileDto;
+    public MessageResponse<ProfileDto> editProfile(ProfileDto profileDto, int id){
+        MessageResponse message = new MessageResponse();
+        if (profileRepository.findById(id).orElse(null)==null){
+            message.setExceptionName("Not found");
+            message.setTechnicalDescription("Profile with your ID not found");
+        }
+        else {
+            Profile profile = REST_MAPPER.update(profileDto, profileRepository.getById(id));
+            profileRepository.save(profile);
+            message.setResponse(profileDto);
+        }
+        return message;
     }
 
     @Loggable
     @Transactional(propagation = Propagation.REQUIRED)
-    public void deleteProfile(int id){
-        profileRepository.deleteById(id);
+    public MessageResponse deleteProfile(int id){
+        MessageResponse message = new MessageResponse();
+        if (profileRepository.findById(id).orElse(null)==null){
+            message.setExceptionName("Not found");
+            message.setTechnicalDescription("Profile with your ID not found");
+        }
+        else {
+            profileRepository.deleteById(id);
+            message.setResponse("Profile deleted");
+        }
+        return message;
     }
 }
